initial commit 00008 - create admin pages
